+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @stackframe/stack-backend
 
+## 2.8.2
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.2
+  - @stackframe/stack-emails@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-emails@2.8.1
+  - @stackframe/stack-shared@2.8.1
+
+## 2.8.0
+
+### Minor Changes
+
+- Various changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.0
+  - @stackframe/stack-emails@2.8.0
+
 ## 2.7.30
 
 ### Patch Changes