+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @stackframe/e2e-tests
 
+## 2.8.2
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.2
+  - @stackframe/js@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- @stackframe/js@2.8.1
+- @stackframe/stack-shared@2.8.1
+
+## 2.8.0
+
+### Minor Changes
+
+- Various changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.0
+  - @stackframe/js@2.8.0
+
 ## 2.7.30
 
 ### Patch Changes