+++ examples/e-commerce/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/e-commerce-demo
 
+## 2.8.2
+
+### Patch Changes
+
+- @stackframe/stack@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- @stackframe/stack@2.8.1
+
+## 2.8.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.8.0
+
 ## 2.7.30
 
 ### Patch Changes